move newConnector
